Renaming: com.horizen -> com.horizen.common
