+++ api/src/rust_jni/merkle_tree.rs
@@ -3,7 +3,7 @@ use crate::ginger_calls::merkle_tree::*;
 use algebra::{CanonicalDeserialize, SemanticallyValid};
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeVerify(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeVerify(
         _env: JNIEnv,
         _path: JObject,
         _height: jint,
@@ -26,14 +26,14 @@ ffi_export!(
         map_to_jboolean_or_throw_exc(
             _env,
             verify_ginger_merkle_path(path, _height as usize, leaf, root),
-            "com/horizen/merkletreenative/MerklePathException",
+            "com/horizen/common/merkletreenative/MerklePathException",
             "Unable to verify MerklePath",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeVerifyWithoutLengthCheck(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeVerifyWithoutLengthCheck(
         _env: JNIEnv,
         _path: JObject,
         _leaf: JObject,
@@ -61,7 +61,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeApply(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeApply(
         _env: JNIEnv,
         _path: JObject,
         _leaf: JObject,
@@ -79,7 +79,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeIsLeftmost(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeIsLeftmost(
         _env: JNIEnv,
         _path: JObject,
     ) -> jboolean {
@@ -91,7 +91,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeIsRightmost(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeIsRightmost(
         _env: JNIEnv,
         _path: JObject,
     ) -> jboolean {
@@ -103,7 +103,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeAreRightLeavesEmpty(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeAreRightLeavesEmpty(
         _env: JNIEnv,
         _path: JObject,
     ) -> jboolean {
@@ -115,7 +115,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeLeafIndex(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeLeafIndex(
         _env: JNIEnv,
         _path: JObject,
     ) -> jlong {
@@ -127,7 +127,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeSerialize(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeSerialize(
         _env: JNIEnv,
         _path: JObject,
     ) -> jbyteArray {
@@ -136,7 +136,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeDeserialize(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeDeserialize(
         _env: JNIEnv,
         _class: JClass,
         _path_bytes: jbyteArray,
@@ -147,14 +147,14 @@ ffi_export!(
             _path_bytes,
             Some(_checked),
             None,
-            "com/horizen/merkletreenative/FieldBasedMerklePath",
-            "com/horizen/merkletreenative/MerklePathException",
+            "com/horizen/common/merkletreenative/FieldBasedMerklePath",
+            "com/horizen/common/merkletreenative/MerklePathException",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_FieldBasedMerklePath_nativeFreeMerklePath(
+    fn Java_com_horizen_common_merkletreenative_FieldBasedMerklePath_nativeFreeMerklePath(
         _env: JNIEnv,
         _path: JObject,
     ) {
@@ -163,7 +163,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeInit(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeInit(
         _env: JNIEnv,
         _class: JClass,
         _height: jint,
@@ -173,15 +173,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             new_ginger_mht(_height as usize, _processing_step as usize),
-            "com/horizen/merkletreenative/BaseMerkleTree",
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/BaseMerkleTree",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to inizialize MerkleTree",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeAppend(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeAppend(
         _env: JNIEnv,
         _tree: JObject,
         _leaf: JObject,
@@ -198,14 +198,14 @@ ffi_export!(
         ok_or_throw_exc!(
             _env,
             append_leaf_to_ginger_mht(tree, leaf),
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to append leaf to MerkleTree"
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeFinalize(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeFinalize(
         _env: JNIEnv,
         _tree: JObject,
     ) -> jobject {
@@ -218,15 +218,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             finalize_ginger_mht(tree),
-            "com/horizen/merkletreenative/BaseMerkleTree",
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/BaseMerkleTree",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to finalize MerkleTree",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeFinalizeInPlace(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeFinalizeInPlace(
         _env: JNIEnv,
         _tree: JObject,
     ) {
@@ -239,14 +239,14 @@ ffi_export!(
         ok_or_throw_exc!(
             _env,
             finalize_ginger_mht_in_place(tree),
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to finalize MerkleTree in place"
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeRoot(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeRoot(
         _env: JNIEnv,
         _tree: JObject,
     ) -> jobject {
@@ -259,15 +259,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             get_ginger_mht_root(tree),
-            "com/horizen/librustsidechains/FieldElement",
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/librustsidechains/FieldElement",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to get MerkleTree root",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeGetMerklePath(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeGetMerklePath(
         _env: JNIEnv,
         _tree: JObject,
         _leaf_index: jlong,
@@ -281,15 +281,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             get_ginger_mht_path(tree, _leaf_index as u64),
-            "com/horizen/merkletreenative/FieldBasedMerklePath",
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/FieldBasedMerklePath",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to get MerklePath",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeGetLeafIndex(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeGetLeafIndex(
         _env: JNIEnv,
         _tree: JObject,
         _leaf: JObject,
@@ -312,7 +312,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeSerialize(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeSerialize(
         _env: JNIEnv,
         _tree: JObject,
     ) -> jbyteArray {
@@ -321,7 +321,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeDeserialize(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeDeserialize(
         _env: JNIEnv,
         _class: JClass,
         _tree_bytes: jbyteArray,
@@ -336,15 +336,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             <GingerMHT as CanonicalDeserialize>::deserialize(obj_bytes.as_slice()),
-            "com/horizen/merkletreenative/BaseMerkleTree",
-            "com/horizen/merkletreenative/MerkleTreeException",
+            "com/horizen/common/merkletreenative/BaseMerkleTree",
+            "com/horizen/common/merkletreenative/MerkleTreeException",
             "Unable to deserialize MerkleTree",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeReset(_env: JNIEnv, _tree: JObject) {
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeReset(_env: JNIEnv, _tree: JObject) {
         let tree = parse_mut_rust_struct_from_jobject::<GingerMHT>(
             &_env,
             _tree,
@@ -356,7 +356,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_merkletreenative_BaseMerkleTree_nativeFreeMerkleTree(
+    fn Java_com_horizen_common_merkletreenative_BaseMerkleTree_nativeFreeMerkleTree(
         _env: JNIEnv,
         _tree: JObject,
     ) {
