+++ api/src/rust_jni/schnorr_signature.rs
@@ -2,7 +2,7 @@ use super::*;
 use crate::ginger_calls::{schnorr_signature::*, serialization::is_valid};
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeGetPublicKeySize(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeGetPublicKeySize(
         _env: JNIEnv,
         _schnorr_public_key_class: JClass,
     ) -> jint {
@@ -11,7 +11,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeSerializePublicKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeSerializePublicKey(
         _env: JNIEnv,
         _schnorr_public_key: JObject,
         _compressed: jboolean,
@@ -26,7 +26,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeDeserializePublicKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeDeserializePublicKey(
         _env: JNIEnv,
         _schnorr_public_key_class: JClass,
         _public_key_bytes: jbyteArray,
@@ -38,14 +38,14 @@ ffi_export!(
             _public_key_bytes,
             Some(_check_public_key),
             Some(_compressed),
-            "com/horizen/schnorrnative/SchnorrPublicKey",
-            "com/horizen/schnorrnative/SchnorrSignatureException",
+            "com/horizen/common/schnorrnative/SchnorrPublicKey",
+            "com/horizen/common/schnorrnative/SchnorrSignatureException",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeFreePublicKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeFreePublicKey(
         _env: JNIEnv,
         _schnorr_public_key: JObject,
     ) {
@@ -54,7 +54,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSecretKey_nativeGetSecretKeySize(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSecretKey_nativeGetSecretKeySize(
         _env: JNIEnv,
         _schnorr_secret_key_class: JClass,
     ) -> jint {
@@ -63,7 +63,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSecretKey_nativeSerializeSecretKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSecretKey_nativeSerializeSecretKey(
         _env: JNIEnv,
         _schnorr_secret_key: JObject,
     ) -> jbyteArray {
@@ -72,7 +72,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSecretKey_nativeDeserializeSecretKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSecretKey_nativeDeserializeSecretKey(
         _env: JNIEnv,
         _schnorr_secret_key_class: JClass,
         _secret_key_bytes: jbyteArray,
@@ -82,14 +82,14 @@ ffi_export!(
             _secret_key_bytes,
             None,
             None,
-            "com/horizen/schnorrnative/SchnorrSecretKey",
-            "com/horizen/schnorrnative/SchnorrSignatureException",
+            "com/horizen/common/schnorrnative/SchnorrSecretKey",
+            "com/horizen/common/schnorrnative/SchnorrSignatureException",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSecretKey_nativeFreeSecretKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSecretKey_nativeFreeSecretKey(
         _env: JNIEnv,
         _schnorr_secret_key: JObject,
     ) {
@@ -99,7 +99,7 @@ ffi_export!(
 
 //Schnorr signature utility functions
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSignature_nativeGetSignatureSize(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSignature_nativeGetSignatureSize(
         _env: JNIEnv,
         _class: JClass,
     ) -> jint {
@@ -108,7 +108,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSignature_nativeSerializeSignature(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSignature_nativeSerializeSignature(
         _env: JNIEnv,
         _schnorr_sig: JObject,
     ) -> jbyteArray {
@@ -117,7 +117,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSignature_nativeDeserializeSignature(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSignature_nativeDeserializeSignature(
         _env: JNIEnv,
         _class: JClass,
         _sig_bytes: jbyteArray,
@@ -128,14 +128,14 @@ ffi_export!(
             _sig_bytes,
             Some(_check_sig),
             None,
-            "com/horizen/schnorrnative/SchnorrSignature",
-            "com/horizen/schnorrnative/SchnorrSignatureException",
+            "com/horizen/common/schnorrnative/SchnorrSignature",
+            "com/horizen/common/schnorrnative/SchnorrSignatureException",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSignature_nativeIsValidSignature(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSignature_nativeIsValidSignature(
         _env: JNIEnv,
         _sig: JObject,
     ) -> jboolean {
@@ -150,7 +150,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSignature_nativefreeSignature(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSignature_nativefreeSignature(
         _env: JNIEnv,
         _sig: JObject,
     ) {
@@ -160,7 +160,7 @@ ffi_export!(
 
 //Schnorr signature functions
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrKeyPair_nativeGenerate(
+    fn Java_com_horizen_common_schnorrnative_SchnorrKeyPair_nativeGenerate(
         _env: JNIEnv,
         // this is the class that owns our
         // static method. Not going to be
@@ -171,17 +171,17 @@ ffi_export!(
         let (pk, sk) = schnorr_generate_key();
 
         let secret_key_object =
-            return_jobject(&_env, sk, "com/horizen/schnorrnative/SchnorrSecretKey");
+            return_jobject(&_env, sk, "com/horizen/common/schnorrnative/SchnorrSecretKey");
         let public_key_object =
-            return_jobject(&_env, pk, "com/horizen/schnorrnative/SchnorrPublicKey");
+            return_jobject(&_env, pk, "com/horizen/common/schnorrnative/SchnorrPublicKey");
 
         let class = _env
-            .find_class("com/horizen/schnorrnative/SchnorrKeyPair")
+            .find_class("com/horizen/common/schnorrnative/SchnorrKeyPair")
             .expect("Should be able to find SchnorrKeyPair class");
 
         let result = _env.new_object(
         class,
-        "(Lcom/horizen/schnorrnative/SchnorrSecretKey;Lcom/horizen/schnorrnative/SchnorrPublicKey;)V",
+        "(Lcom/horizen/common/schnorrnative/SchnorrSecretKey;Lcom/horizen/common/schnorrnative/SchnorrPublicKey;)V",
         &[JValue::Object(secret_key_object), JValue::Object(public_key_object)]
     ).expect("Should be able to create new (SchnorrSecretKey, SchnorrPublicKey) object");
 
@@ -190,7 +190,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrKeyPair_nativeSignMessage(
+    fn Java_com_horizen_common_schnorrnative_SchnorrKeyPair_nativeSignMessage(
         _env: JNIEnv,
         _schnorr_key_pair: JObject,
         _message: JObject,
@@ -200,7 +200,7 @@ ffi_export!(
             .get_field(
                 _schnorr_key_pair,
                 "secretKey",
-                "Lcom/horizen/schnorrnative/SchnorrSecretKey;",
+                "Lcom/horizen/common/schnorrnative/SchnorrSecretKey;",
             )
             .expect("Should be able to get field secretKey")
             .l()
@@ -214,7 +214,7 @@ ffi_export!(
             .get_field(
                 _schnorr_key_pair,
                 "publicKey",
-                "Lcom/horizen/schnorrnative/SchnorrPublicKey;",
+                "Lcom/horizen/common/schnorrnative/SchnorrPublicKey;",
             )
             .expect("Should be able to get field publicKey")
             .l()
@@ -231,15 +231,15 @@ ffi_export!(
         map_to_jobject_or_throw_exc(
             _env,
             schnorr_sign(message, secret_key, public_key),
-            "com/horizen/schnorrnative/SchnorrSignature",
-            "com/horizen/schnorrnative/SchnorrSignatureException",
+            "com/horizen/common/schnorrnative/SchnorrSignature",
+            "com/horizen/common/schnorrnative/SchnorrSignatureException",
             "Unable to sign message",
         )
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeVerifyKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeVerifyKey(
         _env: JNIEnv,
         _public_key: JObject,
     ) -> jboolean {
@@ -255,7 +255,7 @@ ffi_export!(
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrSecretKey_nativeGetPublicKey(
+    fn Java_com_horizen_common_schnorrnative_SchnorrSecretKey_nativeGetPublicKey(
         _env: JNIEnv,
         _secret_key: JObject,
     ) -> jobject {
@@ -264,12 +264,12 @@ ffi_export!(
 
         let pk = schnorr_get_public_key(secret_key);
 
-        return_jobject(&_env, pk, "com/horizen/schnorrnative/SchnorrPublicKey").into_inner()
+        return_jobject(&_env, pk, "com/horizen/common/schnorrnative/SchnorrPublicKey").into_inner()
     }
 );
 
 ffi_export!(
-    fn Java_com_horizen_schnorrnative_SchnorrPublicKey_nativeVerifySignature(
+    fn Java_com_horizen_common_schnorrnative_SchnorrPublicKey_nativeVerifySignature(
         _env: JNIEnv,
         _public_key: JObject,
         _signature: JObject,
@@ -291,7 +291,7 @@ ffi_export!(
         map_to_jboolean_or_throw_exc(
             _env,
             schnorr_verify_signature(message, public_key, signature),
-            "com/horizen/schnorrnative/SchnorrSignatureException",
+            "com/horizen/common/schnorrnative/SchnorrSignatureException",
             "Unable to verify signature",
         )
     }
