+++ build_jar.sh
@@ -10,7 +10,7 @@ cargo test --features "bn_382"
 
 # ######################### Build with tweedle, test and publish ########################
 
-# cargo build -j$(($(nproc)+1)) --release --features "tweedle" --target=x86_64-pc-windows-gnu
+cargo build -j$(($(nproc)+1)) --release --features "tweedle" --target=x86_64-pc-windows-gnu
 cargo build -j$(($(nproc)+1)) --release --features "tweedle" --target=x86_64-unknown-linux-gnu
 
 # ########################
@@ -18,8 +18,8 @@ cargo build -j$(($(nproc)+1)) --release --features "tweedle" --target=x86_64-unk
 mkdir -p jni/src/main/resources/native/linux64
 cp target/x86_64-unknown-linux-gnu/release/libsc_common.so jni/src/main/resources/native/linux64/libsc_common.so
 
-# mkdir -p jni/src/main/resources/native/windows64
-# cp target/x86_64-pc-windows-gnu/release/libsc_common.dll jni/src/main/resources/native/windows64/libsc_common.dll
+mkdir -p jni/src/main/resources/native/windows64
+cp target/x86_64-pc-windows-gnu/release/libsc_common.dll jni/src/main/resources/native/windows64/libsc_common.dll
 
 cd jni
 echo "Building jar"
@@ -38,7 +38,7 @@ cd ..
 ######################### Build with bn382 and publish ########################
 cargo clean
 
-# cargo build -j$(($(nproc)+1)) --release --features "bn_382" --target=x86_64-pc-windows-gnu
+cargo build -j$(($(nproc)+1)) --release --features "bn_382" --target=x86_64-pc-windows-gnu
 cargo build -j$(($(nproc)+1)) --release --features "bn_382" --target=x86_64-unknown-linux-gnu
 
 ########################
@@ -46,8 +46,8 @@ cargo build -j$(($(nproc)+1)) --release --features "bn_382" --target=x86_64-unkn
 mkdir -p jni/src/main/resources/native/linux64
 cp target/x86_64-unknown-linux-gnu/release/libsc_common.so jni/src/main/resources/native/linux64/libsc_common.so
 
-# mkdir -p jni/src/main/resources/native/windows64
-# cp target/x86_64-pc-windows-gnu/release/libsc_common.dll jni/src/main/resources/native/windows64/libsc_common.dll
+mkdir -p jni/src/main/resources/native/windows64
+cp target/x86_64-pc-windows-gnu/release/libsc_common.dll jni/src/main/resources/native/windows64/libsc_common.dll
 
 cd jni
 echo "Building jar"
