+++ jni/src/main/java/com/horizen/common/librustsidechains/FieldElementException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.librustsidechains;
+
+public class FieldElementException extends Exception {
+    public FieldElementException(String message) {
+        super(message);
+    }
+
+    public FieldElementException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
