+++ jni/src/main/java/com/horizen/common/merkletreenative/MerklePathException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.merkletreenative;
+
+public class MerklePathException extends Exception {
+    public MerklePathException(String message) {
+        super(message);
+    }
+
+    public MerklePathException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
