+++ jni/src/main/java/com/horizen/common/merkletreenative/MerkleTreeException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.merkletreenative;
+
+public class MerkleTreeException extends Exception {
+    public MerkleTreeException(String message) {
+        super(message);
+    }
+
+    public MerkleTreeException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
