+++ jni/src/main/java/com/horizen/common/merkletreenative/MerkleTreeLeaf.java
@@ -1,0 +1,13 @@
+package com.horizen.common.merkletreenative;
+
+import com.horizen.common.librustsidechains.FieldElement;
+
+public interface MerkleTreeLeaf extends AutoCloseable {
+    /**
+     * Convert this instance to a FieldElement.
+     * It's caller responsibility to free the FieldElement after using it. 
+     * @return FieldElement representation of this instance
+     * @throws MerkleTreeLeafException if some error occurs during conversion between this and FieldElement
+     */
+    public abstract FieldElement getLeafAsFieldElement() throws MerkleTreeLeafException;
+}
