+++ jni/src/main/java/com/horizen/common/merkletreenative/MerkleTreeLeafException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.merkletreenative;
+
+public class MerkleTreeLeafException extends Exception {
+    public MerkleTreeLeafException(String message) {
+        super(message);
+    }
+
+    public MerkleTreeLeafException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
