+++ jni/src/main/java/com/horizen/common/poseidonnative/PoseidonHashException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.poseidonnative;
+
+public class PoseidonHashException extends Exception {
+    public PoseidonHashException(String message) {
+        super(message);
+    }
+
+    public PoseidonHashException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
