+++ jni/src/main/java/com/horizen/common/schnorrnative/SchnorrSignatureException.java
@@ -1,0 +1,11 @@
+package com.horizen.common.schnorrnative;
+
+public class SchnorrSignatureException extends Exception {
+    public SchnorrSignatureException(String message) {
+        super(message);
+    }
+
+    public SchnorrSignatureException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
