--- jni/src/main/java/com/horizen/librustsidechains/FieldElementException.java
@@ -1,11 +1,0 @@
-package com.horizen.librustsidechains;
-
-public class FieldElementException extends Exception {
-    public FieldElementException(String message) {
-        super(message);
-    }
-
-    public FieldElementException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
