--- jni/src/main/java/com/horizen/merkletreenative/MerklePathException.java
@@ -1,11 +1,0 @@
-package com.horizen.merkletreenative;
-
-public class MerklePathException extends Exception {
-    public MerklePathException(String message) {
-        super(message);
-    }
-
-    public MerklePathException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
