--- jni/src/main/java/com/horizen/merkletreenative/MerkleTreeException.java
@@ -1,11 +1,0 @@
-package com.horizen.merkletreenative;
-
-public class MerkleTreeException extends Exception {
-    public MerkleTreeException(String message) {
-        super(message);
-    }
-
-    public MerkleTreeException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
