--- jni/src/main/java/com/horizen/merkletreenative/MerkleTreeLeaf.java
@@ -1,13 +1,0 @@
-package com.horizen.merkletreenative;
-
-import com.horizen.librustsidechains.FieldElement;
-
-public interface MerkleTreeLeaf extends AutoCloseable {
-    /**
-     * Convert this instance to a FieldElement.
-     * It's caller responsibility to free the FieldElement after using it. 
-     * @return FieldElement representation of this instance
-     * @throws MerkleTreeLeafException if some error occurs during conversion between this and FieldElement
-     */
-    public abstract FieldElement getLeafAsFieldElement() throws MerkleTreeLeafException;
-}
