--- jni/src/main/java/com/horizen/merkletreenative/MerkleTreeLeafException.java
@@ -1,11 +1,0 @@
-package com.horizen.merkletreenative;
-
-public class MerkleTreeLeafException extends Exception {
-    public MerkleTreeLeafException(String message) {
-        super(message);
-    }
-
-    public MerkleTreeLeafException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
