--- jni/src/main/java/com/horizen/poseidonnative/PoseidonHashException.java
@@ -1,11 +1,0 @@
-package com.horizen.poseidonnative;
-
-public class PoseidonHashException extends Exception {
-    public PoseidonHashException(String message) {
-        super(message);
-    }
-
-    public PoseidonHashException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
