--- jni/src/main/java/com/horizen/schnorrnative/SchnorrSignatureException.java
@@ -1,11 +1,0 @@
-package com.horizen.schnorrnative;
-
-public class SchnorrSignatureException extends Exception {
-    public SchnorrSignatureException(String message) {
-        super(message);
-    }
-
-    public SchnorrSignatureException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
