--- jni/src/test/java/com/horizen/TestUtils.java
@@ -1,16 +1,0 @@
-package com.horizen;
-
-import com.google.common.io.BaseEncoding;
-
-public class TestUtils {
-    private TestUtils() {}
-
-    public static byte[] fromHexString(String hex) {
-        return BaseEncoding.base16().lowerCase().decode(hex.toLowerCase());
-    }
-
-    // Get hex string representation of byte array
-    public static String toHexString(byte[] bytes) {
-        return BaseEncoding.base16().lowerCase().encode(bytes);
-    }
-}
