+++ jni/src/test/java/com/horizen/common/TestUtils.java
@@ -1,0 +1,16 @@
+package com.horizen.common;
+
+import com.google.common.io.BaseEncoding;
+
+public class TestUtils {
+    private TestUtils() {}
+
+    public static byte[] fromHexString(String hex) {
+        return BaseEncoding.base16().lowerCase().decode(hex.toLowerCase());
+    }
+
+    // Get hex string representation of byte array
+    public static String toHexString(byte[] bytes) {
+        return BaseEncoding.base16().lowerCase().encode(bytes);
+    }
+}
